+++ src/graphics/screens/post_proc_screen.h
@@ -42,7 +42,7 @@ class PostProcScreen : public IScreen
         std::vector<std::string> *screen_names);
     ~PostProcScreen();
 
-    virtual void update(const float delta_time);
+    virtual void update(const double delta_time);
     virtual void draw(Renderer &renderer, bool lightweight = false);
 };
 }
