+++ src/graphics/screens/scene_test_screen.h
@@ -39,7 +39,7 @@ class SceneTestScreen : public IScreen
         std::vector<std::string> *screen_names);
     ~SceneTestScreen();
 
-    virtual void update(const float delta_time);
+    virtual void update(const double delta_time);
     virtual void draw(Renderer &renderer, bool lightweight = false);
 };
 }
