+++ src/screens/watch_screen.h
@@ -53,6 +53,6 @@ class WatchScreen : public IScreen
     ~WatchScreen();
 
     virtual void draw(Renderer &renderer, bool lightweight = false);
-    void update(const float delta_time);
+    void update(const double delta_time);
 };
 }
